+++ src/components/Home/Carousel/index.tsx
@@ -1,0 +1,5 @@
+import { FC } from 'react';
+
+export const Carousel: FC = () => {
+  return <></>;
+};
