+++ src/components/Home/index.tsx
@@ -1,0 +1,8 @@
+import { Carousel } from './Carousel';
+import { FC } from 'react';
+
+const Home: FC = () => {
+  return <Carousel />;
+};
+
+export default Home;
