+++ src/pages/index.tsx
@@ -1,7 +1,10 @@
-export default function Home() {
+import Home from '@/components/Home';
+
+export default function HomePage() {
   return (
     <>
       {/* TODO: Remove below div and add component for home page*/}
+      <Home />
       <div style={{ height: '300vh' }}></div>
     </>
   );
